requirements.txt
Django>=4.2,<6.0
python-dotenv>=1.0.0
gunicorn>=20.0.0
whitenoise>=6.0.0
requests>=2.25.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
google-generativeai>=0.3.0
supabase>=2.0.0
